Move ApiSessionCreate class to Functions\Company
