+++ test/Intacct/Functions/ApiSessionCreateTest.php
@@ -14,7 +14,7 @@
  * permissions and limitations under the License.
  */
 
-namespace Intacct\Functions;
+namespace Intacct\Functions\Company;
 
 use Intacct\Xml\XMLWriter;
